Update (Problem-1491)-Average Salary Excluding the Minimum and Maximum Salary.cpp

+++ (Problem-1491)-Average Salary Excluding the Minimum and Maximum Salary.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     double average(vector<int>& salary) {
-        int salsum = 0;
-        int salavg = 0;
+        double salsum = 0;
+        double salavg = 0;
         sort(salary.begin(),salary.end());
         salary.pop_back();       
         for(int i = 1; i < salary.size(); i++){
